Refactor pakietów
