+++ src/main/java/com/example/service_to_booking_visit/repository/CalendarRepository.java
@@ -1,7 +1,7 @@
-package repository;
+package com.example.service_to_booking_visit.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import persistance.Calendar;
+import com.example.service_to_booking_visit.persistance.Calendar;
 
 public interface CalendarRepository extends JpaRepository<Calendar, Long> {
