+++ src/main/java/com/example/service_to_booking_visit/repository/CompanyRepository.java
@@ -1,7 +1,7 @@
-package repository;
+package com.example.service_to_booking_visit.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import persistance.Company;
+import com.example.service_to_booking_visit.persistance.Company;
 
 public interface CompanyRepository extends JpaRepository<Company, Long> {
 }
