+++ src/main/java/com/example/service_to_booking_visit/repository/ReservationRepository.java
@@ -1,7 +1,7 @@
-package repository;
+package com.example.service_to_booking_visit.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import persistance.Reservation;
+import com.example.service_to_booking_visit.persistance.Reservation;
 
 public interface ReservationRepository extends JpaRepository<Reservation, Long> {
 }
