+++ src/main/java/com/example/service_to_booking_visit/repository/ServiceRepository.java
@@ -1,7 +1,7 @@
-package repository;
+package com.example.service_to_booking_visit.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import persistance.Service;
+import com.example.service_to_booking_visit.persistance.Service;
 
 public interface ServiceRepository extends JpaRepository<Service, Long> {
