+++ src/main/java/com/example/service_to_booking_visit/service/CalendarService.java
@@ -1,17 +1,17 @@
-package service;
+package com.example.service_to_booking_visit.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import persistance.Calendar;
-import persistance.Client;
-import repository.CalendarRepository;
+import com.example.service_to_booking_visit.persistance.Calendar;
+import com.example.service_to_booking_visit.persistance.Client;
+import com.example.service_to_booking_visit.repository.CalendarRepository;
 
-import javax.persistence.Entity;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class CalendarService {
+
     private final CalendarRepository calendarRepository;
 
     public Calendar findById(Long id) {
