+++ src/main/java/com/example/service_to_booking_visit/service/ClientService.java
@@ -1,9 +1,9 @@
-package service;
+package com.example.service_to_booking_visit.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import persistance.Client;
-import repository.ClientRepository;
+import com.example.service_to_booking_visit.persistance.Client;
+import com.example.service_to_booking_visit.repository.ClientRepository;
 
 import java.util.List;
