+++ src/main/java/com/example/service_to_booking_visit/web/CompanyController.java
@@ -1,11 +1,11 @@
-package web;
+package com.example.service_to_booking_visit.web;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import persistance.Client;
-import persistance.Company;
-import service.CompanyService;
+import com.example.service_to_booking_visit.persistance.Client;
+import com.example.service_to_booking_visit.persistance.Company;
+import com.example.service_to_booking_visit.service.CompanyService;
 
 import java.util.List;
