+++ src/main/java/com/example/service_to_booking_visit/web/ServiceController.java
@@ -1,11 +1,11 @@
-package web;
+package com.example.service_to_booking_visit.web;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import persistance.Client;
-import persistance.Service;
-import service.ServiceService;
+import com.example.service_to_booking_visit.persistance.Client;
+import com.example.service_to_booking_visit.persistance.Service;
+import com.example.service_to_booking_visit.service.ServiceService;
 
 import java.util.List;
