+++ src/main/resources/application.properties
@@ -1,1 +1,6 @@
+spring.datasource.url=jdbc:mysql://localhost:3306/booksy?serverTimezone=UTC&createDatabaseIfNotExist=true
+spring.datasource.username=root
+spring.datasource.password=root
 
+spring.jpa.generate-ddl=true
+spring.jpa.hibernate.ddl-auto=update
